feat(config)!: new 'precision' config to limit similarity decimal digits (default=3)
having a fixed precision prevents subsequent runs of the plugin to
generate different similarity values with no actual difference, yet
pointlessly editing previously generated files in _data/related_posts dir.
This may come annoying for setups that looks at those files and require
them to be committed automatically (e.g. CD pipelines)

+++ sql/supabase/create.sql
@@ -27,7 +27,7 @@ returns table (
   most_recent_edit timestamp with time zone,
   url text,
   date text,
-  similarity float
+  similarity numeric
 )
 language plpgsql
 security definer
